+++ Assets/Scripts/Turrets/TurretBullet.cs
@@ -5,10 +5,24 @@ using UnityEngine;
 public class TurretBullet : BaseTurret {
 
     public GameObject bulletPrefab;
+    Transform bulletOrigin;
 
     // Use this for initialization
-    void Start () {
-        Init();
-        SetAttackBehavior(new AttackBullet(bulletPrefab, transform.GetChild(0).GetChild(0)));
+    protected new void Start () {
+        base.Start();
+
+        if (! bulletPrefab) Debug.LogError("No bullet prefab given!");
+        bulletOrigin = transform.GetChild(0).GetChild(0);
+    }
+
+    protected override void Attack()
+    {
+        GameObject bullet = Instantiate(bulletPrefab, bulletOrigin.position, bulletOrigin.rotation, transform);
+        bullet.GetComponent<Bullet>().Seek(target, this);
+    }
+
+    public void RemoveTarget()
+    {
+        target = null;
     }
 }